move node and traverse! into their own files
