+++ src/cwbvh/mod.rs
@@ -1,11 +1,13 @@
 pub mod builder;
 pub mod bvh2_to_cwbvh;
+pub mod node;
 pub mod simd;
+pub mod traverse_macro;
 
 use std::collections::{HashMap, HashSet};
 
-use bytemuck::{Pod, Zeroable};
-use glam::{uvec2, vec3a, UVec2, UVec3, Vec3, Vec3A};
+use glam::{uvec2, UVec2, UVec3, Vec3A};
+use node::CwBvhNode;
 
 use crate::{
     aabb::Aabb,
@@ -13,290 +15,9 @@ use crate::{
     Boundable, PerComponent,
 };
 
-/// Traverse the BVH with custom node and primitive intersections.
-/// I really didn't want to use a macro but it seems like everything else using closures/yielding is slower given
-/// both generic node and primitive traversal.
-///
-/// # Parameters
-/// - `$cwbvh`: `&CwBvh` The complete bounding volume hierarchy to traverse.
-/// - `$node`: `&CwBvhNode` The current node in the BVH that is being traversed.
-/// - `$state`: `Traversal` Mutable traversal state.
-/// - `$node_intersection`: An expression that is executed at for each node intersection during traversal.
-///     It should test for intersection against use the current `node`, making use of `state.oct_inv4` u32.
-///     It should return a u32 `hitmask` the node children hitmask corresponding to which nodes were intersected.
-/// - `$primitive_intersection`: A code block that is executed for each primitive intersection.
-///     It should read the current `state.primitive_id` u32. This is the index into the primitive indices for the
-///     current primitive to be tested. Use `break` to halt traversal.
-///
-/// # Example: Closest hit ray traversal
-/// ```
-/// use obvhs::{
-///     cwbvh::{builder::build_cwbvh_from_tris, CwBvhNode},
-///     ray::{Ray, RayHit},
-///     test_util::geometry::{icosphere, PLANE},
-///     triangle::Triangle,
-///     BvhBuildParams,
-///     traverse,
-/// };
-/// use glam::*;
-///
-/// let mut tris: Vec<Triangle> = Vec::new();
-/// tris.extend(icosphere(1));
-/// tris.extend(PLANE);
-///
-/// let ray = Ray::new_inf(vec3a(0.1, 0.1, 4.0), vec3a(0.0, 0.0, -1.0));
-///
-/// let bvh = build_cwbvh_from_tris(&tris, BvhBuildParams::medium_build(), &mut 0.0);
-/// let mut hit = RayHit::none();
-/// let mut traverse_ray = ray.clone();
-/// let mut state = bvh.new_traversal(ray.direction);
-/// let mut node;
-/// traverse!(bvh, node, state,
-///     // Node intersection:
-///     CwBvhNode::intersect(node, &traverse_ray, state.oct_inv4),
-///     // Primitive intersection:
-///     {
-///         let t = tris[bvh.primitive_indices[state.primitive_id as usize] as usize].intersect(&traverse_ray);
-///         if t < traverse_ray.tmax {
-///             hit.primitive_id = state.primitive_id;
-///             hit.t = t;
-///             traverse_ray.tmax = t;
-///         }
-///     }
-/// );
-///
-/// let did_hit = hit.t < ray.tmax;
-/// assert!(did_hit);
-/// assert!(bvh.primitive_indices[hit.primitive_id as usize] == 62);
-/// ```
-#[macro_export]
-macro_rules! traverse {
-    ($cwbvh:expr, $node:expr, $state:expr, $node_intersection:expr, $primitive_intersection:expr) => {{
-        loop {
-            // While the primitive group is not empty
-            while $state.primitive_group.y != 0 {
-                let local_primitive_index = $crate::cwbvh::firstbithigh($state.primitive_group.y);
-
-                // Remove primitive from current_group
-                $state.primitive_group.y &= !(1u32 << local_primitive_index);
-
-                $state.primitive_id = $state.primitive_group.x + local_primitive_index;
-                $primitive_intersection
-            }
-            $state.primitive_group = UVec2::ZERO;
-
-            // If there's remaining nodes in the current group to check
-            if $state.current_group.y & 0xff000000 != 0 {
-                let hits_imask = $state.current_group.y;
-
-                let child_index_offset = $crate::cwbvh::firstbithigh(hits_imask);
-                let child_index_base = $state.current_group.x;
-
-                // Remove node from current_group
-                $state.current_group.y &= !(1u32 << child_index_offset);
-
-                // If the node group is not yet empty, push it on the stack
-                if $state.current_group.y & 0xff000000 != 0 {
-                    $state.stack.push($state.current_group);
-                }
-
-                let slot_index = (child_index_offset - 24) ^ ($state.oct_inv4 & 0xff);
-                let relative_index = (hits_imask & !(0xffffffffu32 << slot_index)).count_ones();
-
-                let child_node_index = child_index_base + relative_index;
-
-                $node = &$cwbvh.nodes[child_node_index as usize];
-
-                $state.hitmask = $node_intersection;
-
-                $state.current_group.x = $node.child_base_idx;
-                $state.primitive_group.x = $node.primitive_base_idx;
-
-                $state.current_group.y = (&$state.hitmask & 0xff000000u32) | ($node.imask as u32);
-                $state.primitive_group.y = &$state.hitmask & 0x00ffffffu32;
-            } else {
-                $state.primitive_group = $state.current_group;
-                $state.current_group = UVec2::ZERO;
-            }
-
-            // If there's no remaining nodes in the current group to check, pop it off the stack.
-            if $state.primitive_group.y == 0 && ($state.current_group.y & 0xff000000) == 0 {
-                // If the stack is empty, end traversal.
-                if $state.stack.is_empty() {
-                    $state.current_group.y = 0;
-                    break;
-                }
-
-                $state.current_group = $state.stack.pop_fast();
-            }
-        }
-    }};
-}
-
-//--------------------------------------------
-
 pub const BRANCHING: usize = 8;
-/// A Compressed Wide BVH8 Node
-#[derive(Debug, Clone, Copy, Default, PartialEq)]
-#[repr(C)]
-pub struct CwBvhNode {
-    /// Min point
-    pub p: Vec3,
-    /// Exponent of child bounding box compression
-    pub e: [u8; 3],
-    /// Indicates which children are internal nodes
-    pub imask: u8,
-    pub child_base_idx: u32,
-    pub primitive_base_idx: u32,
-    pub child_meta: [u8; 8],
 
-    // Note: deviation from the paper: the min&max are interleaved here.
-    pub child_min_x: [u8; 8],
-    pub child_max_x: [u8; 8],
-    pub child_min_y: [u8; 8],
-    pub child_max_y: [u8; 8],
-    pub child_min_z: [u8; 8],
-    pub child_max_z: [u8; 8],
-}
-
-unsafe impl Pod for CwBvhNode {}
-unsafe impl Zeroable for CwBvhNode {}
-
-#[inline(always)]
-#[allow(dead_code)]
-pub fn extract_byte(x: u32, b: u32) -> u32 {
-    (x >> (b * 8)) & 0xFFu32
-}
-
-#[inline(always)]
-#[allow(dead_code)]
-pub fn extract_byte64(x: u64, b: usize) -> u32 {
-    ((x >> (b * 8)) as u32) & 0xFFu32
-}
-
-#[inline(always)]
-pub fn extract_u32(data: &[u8; 8], second: bool) -> u32 {
-    unsafe { *(data.as_ptr().add(if second { 0 } else { 4 }) as *const u32) }
-}
-
-const EPSILON: f32 = 0.0001;
 const INVALID: u32 = u32::MAX;
-
-impl CwBvhNode {
-    #[inline(always)]
-    pub fn intersect(&self, ray: &Ray, oct_inv4: u32) -> u32 {
-        let adjusted_ray_dir_inv = vec3a(
-            f32::from_bits((self.e[0] as u32) << 23),
-            f32::from_bits((self.e[1] as u32) << 23),
-            f32::from_bits((self.e[2] as u32) << 23),
-        ) * ray.inv_direction;
-        let adjusted_ray_origin = (Vec3A::from(self.p) - ray.origin) * ray.inv_direction;
-
-        let mut hit_mask = 0;
-
-        let rdx = ray.direction.x < 0.0;
-        let rdy = ray.direction.y < 0.0;
-        let rdz = ray.direction.z < 0.0;
-
-        // [unroll]
-        for i in 0..2 {
-            let meta4 = extract_u32(&self.child_meta, i == 0);
-            let is_inner4 = (meta4 & (meta4 << 1)) & 0x10101010;
-            let inner_mask4 = (is_inner4 >> 4) * 0xffu32;
-            let bit_index4 = (meta4 ^ (oct_inv4 & inner_mask4)) & 0x1f1f1f1f;
-            let child_bits4 = (meta4 >> 5) & 0x07070707;
-
-            // [unroll]
-            for j in 0..4 {
-                let ch = i * 4 + j;
-
-                let q_lo_x = self.child_min_x[ch];
-                let q_lo_y = self.child_min_y[ch];
-                let q_lo_z = self.child_min_z[ch];
-
-                let q_hi_x = self.child_max_x[ch];
-                let q_hi_y = self.child_max_y[ch];
-                let q_hi_z = self.child_max_z[ch];
-
-                let x_min = if rdx { q_hi_x } else { q_lo_x };
-                let x_max = if rdx { q_lo_x } else { q_hi_x };
-                let y_min = if rdy { q_hi_y } else { q_lo_y };
-                let y_max = if rdy { q_lo_y } else { q_hi_y };
-                let z_min = if rdz { q_hi_z } else { q_lo_z };
-                let z_max = if rdz { q_lo_z } else { q_hi_z };
-
-                let mut tmin3 = vec3a(x_min as f32, y_min as f32, z_min as f32);
-                let mut tmax3 = vec3a(x_max as f32, y_max as f32, z_max as f32);
-
-                // Account for grid origin and scale
-                tmin3 = tmin3 * adjusted_ray_dir_inv + adjusted_ray_origin;
-                tmax3 = tmax3 * adjusted_ray_dir_inv + adjusted_ray_origin;
-
-                let tmin = tmin3.x.max(tmin3.y).max(tmin3.z).max(EPSILON); //ray.tmin?
-                let tmax = tmax3.x.min(tmax3.y).min(tmax3.z).min(ray.tmax);
-
-                let intersected = tmin <= tmax;
-                if intersected {
-                    let child_bits = extract_byte(child_bits4, j as u32);
-                    let bit_index = extract_byte(bit_index4, j as u32);
-
-                    hit_mask |= child_bits << bit_index;
-                }
-            }
-        }
-
-        hit_mask
-    }
-
-    #[inline(always)]
-    pub fn intersect_aabb(&self, aabb: &Aabb, oct_inv4: u32) -> u32 {
-        let extent = vec3a(
-            f32::from_bits((self.e[0] as u32) << 23),
-            f32::from_bits((self.e[1] as u32) << 23),
-            f32::from_bits((self.e[2] as u32) << 23),
-        );
-        let extent_rcp = 1.0 / extent;
-        let p = Vec3A::from(self.p);
-
-        // Transform the query aabb into the node's local space
-        let adjusted_aabb = Aabb::new((aabb.min - p) * extent_rcp, (aabb.max - p) * extent_rcp);
-
-        let mut hit_mask = 0;
-
-        for i in 0..2 {
-            let meta4 = extract_u32(&self.child_meta, i == 0);
-            let is_inner4 = (meta4 & (meta4 << 1)) & 0x10101010;
-            let inner_mask4 = (is_inner4 >> 4) * 0xffu32;
-            let bit_index4 = (meta4 ^ (oct_inv4 & inner_mask4)) & 0x1f1f1f1f;
-            let child_bits4 = (meta4 >> 5) & 0x07070707;
-
-            for j in 0..4 {
-                let ch = i * 4 + j;
-                let child_aabb = Aabb::new(
-                    vec3a(
-                        self.child_min_x[ch] as f32,
-                        self.child_min_y[ch] as f32,
-                        self.child_min_z[ch] as f32,
-                    ),
-                    vec3a(
-                        self.child_max_x[ch] as f32,
-                        self.child_max_y[ch] as f32,
-                        self.child_max_z[ch] as f32,
-                    ),
-                );
-
-                if child_aabb.aabb_intersect(&adjusted_aabb) {
-                    let child_bits = extract_byte(child_bits4, j as u32);
-                    let bit_index = extract_byte(bit_index4, j as u32);
-
-                    hit_mask |= child_bits << bit_index;
-                }
-            }
-        }
-
-        hit_mask
-    }
-}
 
 /// A Compressed Wide BVH8
 #[derive(Clone, Default, PartialEq)]
@@ -402,7 +123,7 @@ impl Default for Traversal {
 
 impl Traversal {
     #[inline(always)]
-    /// Reinitialize traversal state with new ray.
+    /// Reinitialize traversal state with new traversal direction.
     pub fn reinit(&mut self, traversal_direction: Vec3A) {
         self.stack.clear();
         self.current_group = uvec2(0, 0x80000000);
@@ -437,6 +158,7 @@ impl CwBvh {
     }
 
     #[inline(always)]
+    /// traversal_direction is used to determine the order of bvh node child traversal. This would typically be the ray direction.
     pub fn new_traversal(&self, traversal_direction: Vec3A) -> Traversal {
         //  BVH8's tend to be shallow. A stack of 32 would be very deep even for a large scene with no tlas.
         let stack = TraversalStack32::default();
@@ -466,35 +188,14 @@ impl CwBvh {
         hit: &mut RayHit,
         mut intersection_fn: F,
     ) -> bool {
-        //let mut state = self.new_traversal(ray);
-        //while self.traverse_dynamic(&mut state, hit, &mut intersection_fn) {}
-        //hit.t < ray.tmax // Note this is valid since traverse_dynamic does not mutate the ray
-
         let mut traverse_ray = ray.clone();
-
         let mut state = self.new_traversal(ray.direction);
         let mut node;
-        traverse!(
+        crate::traverse!(
             self,
             node,
             state,
-            {
-                #[cfg(all(
-                    any(target_arch = "x86", target_arch = "x86_64"),
-                    target_feature = "sse2"
-                ))]
-                {
-                    CwBvhNode::intersect_simd(node, &traverse_ray, state.oct_inv4)
-                }
-
-                #[cfg(not(all(
-                    any(target_arch = "x86", target_arch = "x86_64"),
-                    target_feature = "sse2"
-                )))]
-                {
-                    CwBvhNode::intersect(node, &traverse_ray, state.oct_inv4)
-                }
-            },
+            CwBvhNode::intersect_ray(node, &traverse_ray, state.oct_inv4),
             {
                 let t = intersection_fn(&traverse_ray, state.primitive_id as usize);
                 if t < traverse_ray.tmax {
@@ -505,7 +206,11 @@ impl CwBvh {
             }
         );
 
-        hit.t < ray.tmax
+        // Alternatively (performance seems slightly slower):
+        // let mut state = self.new_ray_traversal(ray);
+        // while self.traverse_dynamic(&mut state, hit, &mut intersection_fn) {}
+
+        hit.t < ray.tmax // Note this is valid since traverse_dynamic does not mutate the ray
     }
 
     /// Traverse the BVH
@@ -564,17 +269,7 @@ impl CwBvh {
 
                 let node = &self.nodes[child_node_index as usize];
 
-                #[cfg(all(
-                    any(target_arch = "x86", target_arch = "x86_64"),
-                    target_feature = "sse2"
-                ))]
-                let hitmask = node.intersect_simd(&state.ray, state.oct_inv4);
-
-                #[cfg(not(all(
-                    any(target_arch = "x86", target_arch = "x86_64"),
-                    target_feature = "sse2"
-                )))]
-                let hitmask = node.intersect(&ray, oct_inv4);
+                let hitmask = node.intersect_ray(&state.ray, state.oct_inv4);
 
                 state.current_group.x = node.child_base_idx;
                 state.primitive_group.x = node.primitive_base_idx;
@@ -652,17 +347,7 @@ impl CwBvh {
 
                 let node = &bvh.nodes[child_node_index as usize];
 
-                #[cfg(all(
-                    any(target_arch = "x86", target_arch = "x86_64"),
-                    target_feature = "sse2"
-                ))]
-                let hitmask = node.intersect_simd(&ray, oct_inv4);
-
-                #[cfg(not(all(
-                    any(target_arch = "x86", target_arch = "x86_64"),
-                    target_feature = "sse2"
-                )))]
-                let hitmask = node.intersect(&ray, oct_inv4);
+                let hitmask = node.intersect_ray(&ray, oct_inv4);
 
                 current_group.x = node.child_base_idx;
                 primitive_group.x = node.primitive_base_idx;
